test generator working
Automatically build a test data set from the existing test cases.
I found that one test appears broken on Windows, need to test on Linux
to confirm. Also, stdout testing seem unreliable. We'll have to use
an intermediary file I think. Or add an override that write to a buffer?